Clean up playwright tests
